Move point base entity to separate module
