+++ homeassistant/components/point/binary_sensor.py
@@ -16,8 +16,8 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import MinutPointEntity
 from .const import DOMAIN as POINT_DOMAIN, POINT_DISCOVERY_NEW, SIGNAL_WEBHOOK
+from .entity import MinutPointEntity
 
 _LOGGER = logging.getLogger(__name__)
